Extract cuboid property handling into mixin
In preparation for adding properties to other classes, let's extract
the missing method handling for properties into a mixin.
