+++ lib/physical.rb
@@ -4,6 +4,7 @@ require "money"
 require "measured/density"
 require "physical/types"
 require "physical/version"
+require "physical/property_readers"
 require "physical/cuboid"
 require "physical/box"
 require "physical/package"
